requirements.txt
anthropic==0.29.0
feedparser==6.0.10
python-dotenv==1.0.0
tavily-python==0.1.9
requests==2.28.2
pydantic>=2.0.0
beautifulsoup4==4.10.0